+++ R/utils.R
@@ -657,8 +657,3 @@ reinitialize <- function(x) {
     }
   }
 }
-
-chunk_time <- function(x, chunks) {
-  ny <- length(x)
-  split(x, rep(seq_len(chunks), each = ceiling(ny/chunks))[seq_along(x)])
-}